+++ hypervisor/arch/x86/sgx.c
@@ -59,7 +59,7 @@ static int32_t partition_epc(void)
 	uint64_t psec_addr = 0UL, psec_size = 0UL;
 	uint64_t free_size = 0UL, alloc_size;
 	struct acrn_vm_config *vm_config = get_vm_config(vm_id);
-	uint64_t vm_request_size = vm_config->epc.size;
+	uint64_t vm_request_size = vm_config->arch.epc.size;
 	int32_t ret = 0;
 
 	while (psec_id < MAX_EPC_SECTIONS) {
@@ -70,7 +70,7 @@ static int32_t partition_epc(void)
 			}
 			mid = 0U;
 			vm_config = get_vm_config(vm_id);
-			vm_request_size = vm_config->epc.size;
+			vm_request_size = vm_config->arch.epc.size;
 		} else {
 			if (free_size == 0UL) {
 				ret = get_epc_section(psec_id, &psec_addr, &psec_size);
@@ -83,7 +83,7 @@ static int32_t partition_epc(void)
 			alloc_size = min(vm_request_size, free_size);
 			vm_epc_maps[mid][vm_id].size = alloc_size;
 			vm_epc_maps[mid][vm_id].hpa = psec_addr + psec_size - free_size;
-			vm_epc_maps[mid][vm_id].gpa = vm_config->epc.base + vm_config->epc.size - vm_request_size;
+			vm_epc_maps[mid][vm_id].gpa = vm_config->arch.epc.base + vm_config->arch.epc.size - vm_request_size;
 			vm_request_size -= alloc_size;
 			free_size -= alloc_size;
 			mid++;
