+++ misc/config_tools/data/generic_board/generic_code/hybrid/pci_dev.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <asm/vm_config.h>
+#include <vm_config.h>
 #include <vpci.h>
 #include <asm/mmu.h>
 #include <asm/page.h>
