+++ misc/config_tools/data/generic_board/generic_code/hybrid/pt_intx.c
@@ -4,5 +4,5 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <asm/vm_config.h>
+#include <vm_config.h>
 struct pt_intx_config vm0_pt_intx[1U] = {};
