+++ misc/config_tools/data/generic_board/generic_code/partitioned/pt_intx.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <asm/vm_config.h>
+#include <vm_config.h>
 struct pt_intx_config vm0_pt_intx[1U] = {};
 
 struct pt_intx_config vm1_pt_intx[1U] = {};
