+++ misc/config_tools/data/generic_board/generic_code/shared/pt_intx.c
@@ -4,4 +4,4 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <asm/vm_config.h>
+#include <vm_config.h>
