+++ misc/config_tools/data/generic_board/generic_code/shared/vm_configurations.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <asm/vm_config.h>
+#include <vm_config.h>
 #include <vuart.h>
 #include <asm/pci_dev.h>
 #include <asm/pgtable.h>
