+++ misc/config_tools/xforms/pt_intx.c.xsl
@@ -18,7 +18,7 @@
     <xsl:value-of select="$license" />
 
     <!-- Included headers -->
-    <xsl:value-of select="acrn:include('asm/vm_config.h')" />
+    <xsl:value-of select="acrn:include('vm_config.h')" />
 
     <xsl:apply-templates select="config-data/acrn-config/vm" />
   </xsl:template>
